updated test with latest structure changes
